feat: added SAT (separating axis theorem) with collision resolution
The physicsworld now completely owns the objects. The objects are
created in main and moved to physicsworld. The physicsworld now has
solve_collisions and resolve_collision; solve_collisions uses the SAT
(separating axis theorem to detect collisions and find the depth and
axis which to move the object(s) in. The resolve_collision function uses
the depth and axis found in solve_collisions (should prob name it
better) to move the object(s)